Rework to use gRPC
This commit reworks the package to use gRPC and protocol buffers for
all the peer-to-peer communications. This simplifies transport encoding
a lot and guarantees us less bandwidth since gossip tends to be a bit
noisy.

What's missing so far is the ability for a package consumer to get info
on the state of the network. What's added is the ability for peers to
cleanly leave the gossip network.

Signed-off-by: David Bond <davidsbond93@gmail.com>