Make peer store implementable & status RPC
This commit modifies the store implementation to use a dedicated peer type
that masks the underlying gRPC type within the gossip protocol. This is
required so that a consumer can implement their own storage system without
having to import generated types.

This will also make it easier to expose events as generated types will not
be required.

This commit also adds a `Status` rpc which can be used to query the state
of the network according to a single peer.

Signed-off-by: David Bond <davidsbond93@gmail.com>