Check peers via other peers
This commit modifies the gossip protocol to periodically check the status
endpoint of a random active peer. If the peer fails to present its current
status we check if the status can be obtained via another peer.

If this second status check fails we'll locally mark the peer as gone
and propagate that change outward to the network.

Signed-off-by: David Bond <davidsbond93@gmail.com>