Add test for failure detection
This commit adds a new test for failure detection by inserting a peer
record into each node's store for a peer that does not exist. Each
peer will eventually attempt to check the status of the fake peer and
mark it with a gone status when it confirms neither peer can reach it.

This required exposing some configuration to tweak the gossip and check
intervals which allows the other tests to run a bit faster.

Also included in this change is modifying the join method so that a
peer marks itself as joined rather than the peer it joins doing so.

Now, each peer manages its own status except in the scenario where one
has been found as offline.

Signed-off-by: David Bond <davidsbond93@gmail.com>